Move libcoap interop test into separate packages.
