+++ californium-tests/californium-interoperability-tests/src/test/java/org/eclipse/californium/interoperability/test/libcoap/gnutls/LibCoapClientGnuTlsInteroperabilityTest.java
@@ -13,10 +13,10 @@
  * Contributors:
  *    Achim Kraus (Bosch.IO GmbH) - initial implementation.
  ******************************************************************************/
-package org.eclipse.californium.interoperability.test.libcoap;
+package org.eclipse.californium.interoperability.test.libcoap.gnutls;
 
-import static org.eclipse.californium.interoperability.test.CredentialslUtil.SERVER_CA_RSA_CERTIFICATE;
 import static org.eclipse.californium.interoperability.test.CredentialslUtil.CLIENT_RSA_CERTIFICATE;
+import static org.eclipse.californium.interoperability.test.CredentialslUtil.SERVER_CA_RSA_CERTIFICATE;
 import static org.eclipse.californium.interoperability.test.ProcessUtil.TIMEOUT_MILLIS;
 import static org.eclipse.californium.interoperability.test.libcoap.LibCoapProcessUtil.REQUEST_TIMEOUT_MILLIS;
 import static org.eclipse.californium.interoperability.test.libcoap.LibCoapProcessUtil.LibCoapAuthenticationMode.CA;
@@ -27,6 +27,7 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeNotNull;
 import static org.junit.Assume.assumeTrue;
 
+import java.io.File;
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
@@ -38,6 +39,7 @@ import org.eclipse.californium.elements.rule.TestNameLoggerRule;
 import org.eclipse.californium.interoperability.test.CaliforniumUtil;
 import org.eclipse.californium.interoperability.test.ConnectorUtil;
 import org.eclipse.californium.interoperability.test.ProcessUtil.ProcessResult;
+import org.eclipse.californium.interoperability.test.libcoap.LibCoapProcessUtil;
 import org.eclipse.californium.interoperability.test.ScandiumUtil;
 import org.eclipse.californium.interoperability.test.ShutdownUtil;
 import org.eclipse.californium.scandium.config.DtlsConfig;
@@ -54,31 +56,52 @@ import org.junit.Rule;
 import org.junit.Test;
 
 /**
- * Test for interoperability with libcoap client using mbedtls.
+ * Test for interoperability with libcoap client using gnutls.
  * 
  * @see LibCoapProcessUtil
  */
-public class LibCoapClientMbedTlsInteroperabilityTest {
+public class LibCoapClientGnuTlsInteroperabilityTest {
 
 	@Rule
 	public TestNameLoggerRule name = new TestNameLoggerRule();
 
 	private static final InetSocketAddress BIND = new InetSocketAddress(InetAddress.getLoopbackAddress(),
 			ScandiumUtil.PORT);
+
+	/**
+	 * Gnutls seems to require a encoded private key with the optional public
+	 * key. <a href="https://tools.ietf.org/html/rfc5915#section-3" target="_blank">RFC 5915 -
+	 * Section 3</a> Unclear, how to achieve that with openssl 1.1.1, seems to
+	 * be the output of openssl 1.0
+	 */
+	private static final String CLIENT_PRIVATE_KEY = "clientPrivateKey.pem";
+	private static final String CLIENT_RSA_PRIVATE_KEY = "clientRsaPrivateKey.pem";
 	private static final String DESTINATION = "127.0.0.1:" + ScandiumUtil.PORT;
 	private static final String DESTINATION_URL = "coaps://" + DESTINATION + "/";
 
 	private static LibCoapProcessUtil processUtil;
 	private static CaliforniumUtil californiumUtil;
+	private static String clientPrivateKey;
+	private static String clientRsaPrivateKey;
 
 	@BeforeClass
 	public static void init() throws IOException, InterruptedException {
 		processUtil = new LibCoapProcessUtil();
-		ProcessResult result = processUtil.prepareLibCoapClientMbedTls(TIMEOUT_MILLIS);
+		ProcessResult result = processUtil.prepareLibCoapClientGnuTls(TIMEOUT_MILLIS);
 		assumeNotNull(result);
-		processUtil.assumeMinVersion("4.3.0");
-		processUtil.assumeMinDtlsVersion("2.16.5");
+		processUtil.assumeMinVersion("4.2.1");
+		processUtil.assumeMinDtlsVersion("3.5.18");
 		californiumUtil = new CaliforniumUtil(false);
+		if (processUtil.compareVersion("4.3.0") >= 0) {
+			File privatekey = new File(CLIENT_PRIVATE_KEY);
+			if (privatekey.isFile() && privatekey.canRead()) {
+				clientPrivateKey = CLIENT_PRIVATE_KEY;
+			}
+			privatekey = new File(CLIENT_RSA_PRIVATE_KEY);
+			if (privatekey.isFile() && privatekey.canRead()) {
+				clientRsaPrivateKey = CLIENT_RSA_PRIVATE_KEY;
+			}
+		}
 	}
 
 	@AfterClass
@@ -86,24 +109,10 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 		ShutdownUtil.shutdown(californiumUtil, processUtil);
 	}
 
-	private DtlsConnectorConfig.Builder builder;
-
 	@Before
-	public void start() {
+	public void start() throws InterruptedException {
 		processUtil.setTag(name.getName());
-		// mbedtls v2.27 still supports only the deprecated MAC calculation.
-		// Ensure/adjust the extension id in mbedtls - include/mbedtls/ssl.h
-		// for compatibility to 53
-		// 
-		// #define MBEDTLS_TLS_EXT_CID                        53
-		// 
-		// For libcoap enable the passive use of CID in src/coap_mbedtls.c,
-		// coap_dtls_new_mbedtls_env, before mbedtls_ssl_set_bio with
-		// 
-		// mbedtls_ssl_set_cid(&m_env->ssl, MBEDTLS_SSL_CID_ENABLED, NULL, 0);
-
-		builder = DtlsConnectorConfig.builder(new Configuration())
-				.set(DtlsConfig.DTLS_SUPPORT_DEPRECATED_CID, true);
+		processUtil.setPrivateKey(clientPrivateKey);
 	}
 
 	@After
@@ -115,7 +124,7 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 	public void testLibCoapClientPskGCM() throws Exception {
 		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_GCM_SHA256;
 		assumeTrue("GCM not support by JCE", cipherSuite.isSupported());
-		californiumUtil.start(BIND, builder, null, cipherSuite);
+		californiumUtil.start(BIND, null, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", PSK, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -125,7 +134,7 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 	@Test
 	public void testLibCoapClientPsk() throws Exception {
 		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_CCM_8;
-		californiumUtil.start(BIND, builder, null, cipherSuite);
+		californiumUtil.start(BIND, null, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", PSK, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -135,8 +144,9 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 	@Test
 	public void testLibCoapClientPskMultiFragment() throws Exception {
 		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_CCM_8;
-		builder.set(DtlsConfig.DTLS_USE_MULTI_HANDSHAKE_MESSAGE_RECORDS, true);
-		californiumUtil.start(BIND,  builder, null, cipherSuite);
+		DtlsConnectorConfig.Builder builder = DtlsConnectorConfig.builder(new Configuration())
+				.set(DtlsConfig.DTLS_USE_MULTI_HANDSHAKE_MESSAGE_RECORDS, true);
+		californiumUtil.start(BIND, builder, null, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", PSK, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -146,7 +156,8 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 	@Test
 	public void testLibCoapClientPskNoSessionId() throws Exception {
 		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_CCM_8;
-		builder.set(DtlsConfig.DTLS_SERVER_USE_SESSION_ID, false);
+		DtlsConnectorConfig.Builder builder = DtlsConnectorConfig.builder(new Configuration())
+				.set(DtlsConfig.DTLS_SERVER_USE_SESSION_ID, false);
 		californiumUtil.start(BIND, builder, null, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", PSK, "Hello, CoAP!", cipherSuite);
@@ -156,9 +167,10 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 
 	@Test
 	public void testLibCoapClientEcdsaGCM() throws Exception {
+		assumeNotNull(clientPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256;
 		assumeTrue("GCM not support by JCE", cipherSuite.isSupported());
-		californiumUtil.start(BIND, builder, null, cipherSuite);
+		californiumUtil.start(BIND, null, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", CHAIN, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -167,8 +179,9 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 
 	@Test
 	public void testLibCoapClientEcdsa() throws Exception {
+		assumeNotNull(clientPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
-		californiumUtil.start(BIND, builder, null, cipherSuite);
+		californiumUtil.start(BIND, null, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", CHAIN, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -177,9 +190,10 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 
 	@Test
 	public void testLibCoapClientEcdsaRsa() throws Exception {
+		assumeNotNull(clientPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
 		californiumUtil.loadCredentials(ConnectorUtil.SERVER_CA_RSA_NAME);
-		californiumUtil.start(BIND, builder, ScandiumUtil.TRUST_ROOT, cipherSuite);
+		californiumUtil.start(BIND, null, ScandiumUtil.TRUST_ROOT, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", CHAIN, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -188,14 +202,16 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 
 	@Test
 	public void testLibCoapClientRsa() throws Exception {
+		assumeNotNull(clientRsaPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256.isSupported()
 				? CipherSuite.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256
-				: CipherSuite.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256;
+				: CipherSuite.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384;
 		assumeTrue(cipherSuite.name() + " not support by JCE", cipherSuite.isSupported());
 
 		californiumUtil.loadCredentials(ConnectorUtil.SERVER_RSA_NAME);
 		californiumUtil.start(BIND, null, ScandiumUtil.TRUST_ROOT, cipherSuite);
 
+		processUtil.setPrivateKey(clientRsaPrivateKey);
 		processUtil.setCertificate(CLIENT_RSA_CERTIFICATE);
 		processUtil.startupClient(DESTINATION_URL + "test", CHAIN, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -204,8 +220,9 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 
 	@Test
 	public void testLibCoapClientEcdsaTrust() throws Exception {
+		assumeNotNull(clientPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
-		californiumUtil.start(BIND, builder, null, cipherSuite);
+		californiumUtil.start(BIND, null, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", TRUST, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
@@ -214,31 +231,32 @@ public class LibCoapClientMbedTlsInteroperabilityTest {
 
 	@Test
 	public void testLibCoapClientEcdsaTrustFails() throws Exception {
+		assumeNotNull(clientPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
-		californiumUtil.start(BIND, builder, null, cipherSuite);
-
+		californiumUtil.start(BIND, null, cipherSuite);
 		processUtil.setTrusts(SERVER_CA_RSA_CERTIFICATE);
 		processUtil.startupClient(DESTINATION_URL + "test", TRUST, "Hello, CoAP!", cipherSuite);
-		connect(null, "X509 - Certificate verification failed");
+		connect(null, "The peer certificate's CA is unknown");
 		californiumUtil.assertAlert(TIMEOUT_MILLIS, new AlertMessage(AlertLevel.FATAL, AlertDescription.UNKNOWN_CA));
 	}
 
 	@Test
 	public void testLibCoapClientEcdsaCaFails() throws Exception {
+		assumeNotNull(clientPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
-		californiumUtil.start(BIND, builder, null, cipherSuite);
-
+		californiumUtil.start(BIND, null, cipherSuite);
 		processUtil.setCa(SERVER_CA_RSA_CERTIFICATE);
 		processUtil.startupClient(DESTINATION_URL + "test", CA, "Hello, CoAP!", cipherSuite);
-		connect(null, "X509 - Certificate verification failed");
+		connect(null, "The peer certificate's CA is unknown");
 		californiumUtil.assertAlert(TIMEOUT_MILLIS, new AlertMessage(AlertLevel.FATAL, AlertDescription.UNKNOWN_CA));
 	}
 
 	@Test
 	public void testLibCoapClientEcdsaRsaTrust() throws Exception {
+		assumeNotNull(clientPrivateKey);
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
 		californiumUtil.loadCredentials(ConnectorUtil.SERVER_CA_RSA_NAME);
-		californiumUtil.start(BIND, builder, ScandiumUtil.TRUST_ROOT, cipherSuite);
+		californiumUtil.start(BIND, null, ScandiumUtil.TRUST_ROOT, cipherSuite);
 
 		processUtil.startupClient(DESTINATION_URL + "test", TRUST, "Hello, CoAP!", cipherSuite);
 		connect("Hello, CoAP!", "Greetings!");
