+++ californium-tests/californium-interoperability-tests/src/test/java/org/eclipse/californium/interoperability/test/libcoap/tinydtls/LibCoapServerTinyDtlsInteroperabilityTest.java
@@ -13,54 +13,57 @@
  * Contributors:
  *    Achim Kraus (Bosch.IO GmbH) - initial implementation.
  ******************************************************************************/
-package org.eclipse.californium.interoperability.test.libcoap;
+package org.eclipse.californium.interoperability.test.libcoap.tinydtls;
 
 import static org.eclipse.californium.interoperability.test.ProcessUtil.TIMEOUT_MILLIS;
 import static org.eclipse.californium.interoperability.test.libcoap.LibCoapProcessUtil.REQUEST_TIMEOUT_MILLIS;
 import static org.eclipse.californium.interoperability.test.libcoap.LibCoapProcessUtil.LibCoapAuthenticationMode.PSK;
 import static org.eclipse.californium.interoperability.test.libcoap.LibCoapProcessUtil.LibCoapAuthenticationMode.RPK;
-import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeNotNull;
 
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 
+import org.eclipse.californium.core.CoapResponse;
+import org.eclipse.californium.core.Utils;
+import org.eclipse.californium.core.coap.CoAP;
+import org.eclipse.californium.core.coap.Request;
 import org.eclipse.californium.elements.auth.PreSharedKeyIdentity;
 import org.eclipse.californium.elements.auth.RawPublicKeyIdentity;
-import org.eclipse.californium.elements.config.CertificateAuthenticationMode;
-import org.eclipse.californium.elements.config.Configuration;
 import org.eclipse.californium.elements.rule.TestNameLoggerRule;
+import org.eclipse.californium.elements.util.StringUtil;
 import org.eclipse.californium.interoperability.test.CaliforniumUtil;
 import org.eclipse.californium.interoperability.test.ProcessUtil.ProcessResult;
+import org.eclipse.californium.interoperability.test.libcoap.LibCoapProcessUtil;
 import org.eclipse.californium.interoperability.test.ScandiumUtil;
 import org.eclipse.californium.interoperability.test.ShutdownUtil;
-import org.eclipse.californium.scandium.config.DtlsConfig;
-import org.eclipse.californium.scandium.config.DtlsConnectorConfig;
 import org.eclipse.californium.scandium.dtls.cipher.CipherSuite;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 
 /**
- * Test for interoperability with libcoap client using tinydtls.
+ * Test for interoperability with libcoap server using tinydtls.
  * 
  * @see LibCoapProcessUtil
  */
-public class LibCoapClientTinyDtlsInteroperabilityTest {
+public class LibCoapServerTinyDtlsInteroperabilityTest {
 
 	@Rule
 	public TestNameLoggerRule name = new TestNameLoggerRule();
 
-	private static final InetSocketAddress BIND = new InetSocketAddress(InetAddress.getLoopbackAddress(),
+	private static final InetSocketAddress BIND = new InetSocketAddress(InetAddress.getLoopbackAddress(), 0);
+	private static final InetSocketAddress DESTINATION = new InetSocketAddress(InetAddress.getLoopbackAddress(),
 			ScandiumUtil.PORT);
-	private static final String DESTINATION = "127.0.0.1:" + ScandiumUtil.PORT;
-	private static final String DESTINATION_URL = "coaps://" + DESTINATION + "/";
+	private static final String ACCEPT = "127.0.0.1:" + ScandiumUtil.PORT;
 
 	private static LibCoapProcessUtil processUtil;
 	private static CaliforniumUtil californiumUtil;
@@ -68,11 +71,11 @@ public class LibCoapClientTinyDtlsInteroperabilityTest {
 	@BeforeClass
 	public static void init() throws IOException, InterruptedException {
 		processUtil = new LibCoapProcessUtil();
-		ProcessResult result = processUtil.prepareLibCoapClientTinyDtls(TIMEOUT_MILLIS);
+		ProcessResult result = processUtil.prepareLibCoapServerTinyDtls(TIMEOUT_MILLIS);
 		assumeNotNull(result);
 		processUtil.assumeMinVersion("4.2.1");
 		processUtil.assumeMinDtlsVersion("0.8.6");
-		californiumUtil = new CaliforniumUtil(false);
+		californiumUtil = new CaliforniumUtil(true);
 	}
 
 	@AfterClass
@@ -91,73 +94,86 @@ public class LibCoapClientTinyDtlsInteroperabilityTest {
 	}
 
 	@Test
-	public void testLibCoapClientTinyDtlsPsk() throws Exception {
+	public void testLibCoapServerPsk() throws Exception {
 		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_CCM_8;
+		processUtil.startupServer(ACCEPT, PSK, cipherSuite);
+
+		californiumUtil.start(BIND, null, cipherSuite);
+		connect(true);
+		californiumUtil.assertPrincipalType(PreSharedKeyIdentity.class);
+	}
+
+	@Test
+	public void testLibCoapServerPsk2FullHandshake() throws Exception {
+		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_CCM_8;
+		processUtil.startupServer(ACCEPT, PSK, cipherSuite);
+
 		californiumUtil.start(BIND, null, cipherSuite);
 
-		processUtil.startupClient(DESTINATION_URL + "test", PSK, "Hello, CoAP!", cipherSuite);
-		connect("Hello, CoAP!", "Greetings!");
-		californiumUtil.assertPrincipalType(PreSharedKeyIdentity.class);
-	}
+		// first handshake
+		Request request = Request.newGet();
+		request.setURI("coaps://" + StringUtil.toString(DESTINATION) + "/time");
+		CoapResponse response = californiumUtil.send(request);
+		assertNotNull(response);
+		assertEquals(CoAP.ResponseCode.CONTENT, response.getCode());
 
-	@Ignore
-	@Test
-	public void testLibCoapClientTinyDtlsPskMultiFragment() throws Exception {
-		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_CCM_8;
-		DtlsConnectorConfig.Builder builder = DtlsConnectorConfig.builder(new Configuration())
-				.set(DtlsConfig.DTLS_USE_MULTI_HANDSHAKE_MESSAGE_RECORDS, true);
-		californiumUtil.start(BIND, builder, null, cipherSuite);
+		// second handshake
+		request = Request.newGet();
+		request.setURI("coaps://" + StringUtil.toString(DESTINATION) + "/time");
+		response = californiumUtil.sendWithFullHandshake(request);
+		assertNotNull(response);
+		assertEquals(CoAP.ResponseCode.CONTENT, response.getCode());
 
-		processUtil.startupClient(DESTINATION_URL + "test", PSK, "Hello, CoAP!", cipherSuite);
-		connect("Hello, CoAP!", "Greetings!");
-		californiumUtil.assertPrincipalType(PreSharedKeyIdentity.class);
-	}
-
-	@Test
-	public void testLibCoapClientTinyDtlsPskNoSessionId() throws Exception {
-		CipherSuite cipherSuite = CipherSuite.TLS_PSK_WITH_AES_128_CCM_8;
-		DtlsConnectorConfig.Builder builder = DtlsConnectorConfig.builder(new Configuration())
-				.set(DtlsConfig.DTLS_SERVER_USE_SESSION_ID, false);
-		californiumUtil.start(BIND, builder, null, cipherSuite);
-
-		processUtil.startupClient(DESTINATION_URL + "test", PSK, "Hello, CoAP!", cipherSuite);
-		connect("Hello, CoAP!", "Greetings!");
+		connect(true);
 		californiumUtil.assertPrincipalType(PreSharedKeyIdentity.class);
 	}
 
 	@Test
-	public void testLibCoapClientTinyDtlsRpk() throws Exception {
+	public void testLibCoapServerRpk() throws Exception {
 		processUtil.assumeMinVersion("4.3.0");
 		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
-		californiumUtil.start(BIND, null, cipherSuite);
+		processUtil.startupServer(ACCEPT, RPK, cipherSuite);
 
-		processUtil.startupClient(DESTINATION_URL + "test", RPK, "Hello, CoAP!", cipherSuite);
-		connect("Hello, CoAP!", "Greetings!", "certificate \\(11\\)", "certificate_verify \\(15\\)");
+		californiumUtil.start(BIND, null, cipherSuite);
+		ProcessResult result = connect(true);
+		assertTrue(result.contains("certificate \\(11\\)"));
+		assertTrue(result.contains("certificate_request \\(13\\)"));
+		assertTrue(result.contains("certificate_verify \\(15\\)"));
 		californiumUtil.assertPrincipalType(RawPublicKeyIdentity.class);
 	}
 
-	@Test
-	public void testLibCoapClientTinyDtlsRpkAnonymousClient() throws Exception {
-		processUtil.assumeMinVersion("4.3.0");
-		CipherSuite cipherSuite = CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CCM_8;
-		Configuration configuration = new Configuration();
-		configuration.set(DtlsConfig.DTLS_CLIENT_AUTHENTICATION_MODE, CertificateAuthenticationMode.NONE);
-		DtlsConnectorConfig.Builder builder = DtlsConnectorConfig.builder(configuration);
-		californiumUtil.start(BIND, builder, null, cipherSuite);
-
-		processUtil.startupClient(DESTINATION_URL + "test", RPK, "Hello, CoAP!", cipherSuite);
-		connect("Hello, CoAP!", "Greetings!", "certificate \\(11\\)");
-		assertNull(californiumUtil.getPrincipal());
-	}
-
-	public ProcessResult connect(String sendMessage, String... patterns) throws Exception {
-		long timeout = REQUEST_TIMEOUT_MILLIS.get();
-		if (patterns != null) {
-			for (String check : patterns) {
-				assertTrue("missing " + check + " (" + timeout + "ms)", processUtil.waitConsole(check, timeout));
+	public ProcessResult connect(boolean success, String... patterns) throws Exception {
+		Request request = Request.newGet();
+		request.setURI("coaps://" + StringUtil.toString(DESTINATION) + "/time");
+		CoapResponse response = californiumUtil.send(request);
+		if (success) {
+			if (response != null) {
+				System.out.println(Utils.prettyPrint(response));
+				assertEquals(CoAP.ResponseCode.CONTENT, response.getCode());
+			} else if (request.getSendError() != null) {
+				fail("error " + request.getSendError());
+			} else if (request.isTimedOut()) {
+				fail("timeout!");
+			} else {
+				fail("unknown cause!");
+			}
+		} else {
+			if (response != null) {
+				System.out.println(Utils.prettyPrint(response));
+				fail("unexpected response!");
+			} else if (request.getSendError() != null) {
+				System.out.println("expected error: " + request.getSendError());
+			} else if (request.isTimedOut()) {
+				fail("timeout!");
+			} else {
+				fail("unknown cause!");
 			}
 		}
-		californiumUtil.assertReceivedData(sendMessage, timeout);
+		if (patterns != null) {
+			for (String check : patterns) {
+				assertTrue("missing " + check, processUtil.waitConsole(check, REQUEST_TIMEOUT_MILLIS.get()));
+			}
+		}
 		return processUtil.stop(TIMEOUT_MILLIS);
 	}
 }
